update : refactor ui components
