+++ ui/layouts/article/LayoutArticleMoreSection.tsx
@@ -1,8 +1,8 @@
 import LayoutArticleSuggestionSection from "./LayoutArticleSuggestionSection";
 import LayoutArticleCommentSection from "./comment/LayoutArticleCommentSection";
-import { ArticleModel } from "../../base/data/models/ArticleModel";
-import useLazyScrollerHook from "../../app/hooks/LazyScrollerHook";
 import LoadingIndicator from "../../components/placeholder/LoadingIndicator";
+import useLazyScrollerHook from "../../../app/hooks/LazyScrollerHook";
+import { ArticleModel } from "../../../base/data/models/ArticleModel";
 
 // Supposed to serve as container of comment and suggestion section
 function LayoutArticleMoreSection({ article }: { article: ArticleModel }) {
