+++ ui/layouts/article/pages/LayoutArticlePageSlug.tsx
@@ -8,6 +8,8 @@ import {
   MdStar,
   MdTrendingUp,
 } from "react-icons/md";
+import { fbArticleContentGet, fbArticleUpdateView } from "../../../../app/services/ArticleService";
+import { ArticleModel } from "../../../../base/data/models/ArticleModel";
 import ArticleSectionAction from "../../../components/article/ArticleActions";
 import ArticleProgressBar from "../../../components/article/ArticleProgressBar";
 import MainContainer from "../../../components/main/MainContainer";
@@ -18,11 +20,6 @@ import MainUserPopup from "../../../components/main/MainPostUserPopup";
 import MobileHeader from "../../../components/main/MobileHeader";
 import LoadingIndicator from "../../../components/placeholder/LoadingIndicator";
 import UserHeader from "../../../components/user/UserHeader";
-import { ArticleModel } from "../../../base/data/models/ArticleModel";
-import {
-  fbArticleContentGet,
-  fbArticleUpdateView,
-} from "../../../app/services/ArticleService";
 import LayoutArticleMoreSection from "../LayoutArticleMoreSection";
 
 function LayoutArticlePageSlug({
