+++ ui/layouts/article/pages/LayoutArticlePageWrite.tsx
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router";
 import React from "react";
+import { WritingPanelProvider } from "../../../../app/contexts/writingPanel/WritingPanelProvider";
 import MainContainer from "../../../components/main/MainContainer";
 import MobileHeader from "../../../components/main/MobileHeader";
 import WritingPanel from "../../../components/write/WritingPanel";
-import { WritingPanelProvider } from "../../../app/contexts/writingPanel/WritingPanelProvider";
 
 function LayoutArticlePageWrite() {
   const router = useRouter();
