+++ ui/layouts/index/pages/LayoutIndexPage.tsx
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router";
 import React from "react";
 import { MdEdit } from "react-icons/md";
+import { APP_NAME } from "../../../../app/helpers/Constants";
 import MainContainer from "../../../components/main/MainContainer";
 import MobileHeader from "../../../components/main/MobileHeader";
-import { APP_NAME } from "../../../app/helpers/Constants";
 import LayoutIndexPostSection from "../LayoutIndexPostSection";
 import LayoutIndexTabFilter from "../LayoutIndexTabFilter";
